backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary>=2.9.0
pydantic==2.11.9
pydantic-settings==2.1.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
httpx==0.25.2
python-dotenv==1.0.0
authlib==1.2.1
itsdangerous==2.1.2
mypy==1.7.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
asyncpg==0.29.0
email-validator==2.1.1
pyjwt[crypto]==2.8.0
cryptography==41.0.7
